refactor: decoupled demos from error_correction_algorithms
