+++ CMakeLists.txt
@@ -49,6 +49,7 @@ set(SRC_DIRS
     demos/dynamic_programming
     demos/expression_evaluation
     demos/advanced_heaps
+    demos/error_correction_algorithms
     features/telemetry
     features/serialization
     features/dp_visualizer
