+++ demos/error_correction_algorithms/hamming_receiver_demo.c
