+++ demos/error_correction_algorithms/lrc_reciever_demo.c
