+++ src/error_correction_algorithms/crc.c
@@ -93,57 +93,3 @@ int crc_verify(const char* codeword, const char* generator, char* remainder_out)
     }
     return 1; // No error
 }
-
-void crc_demo(void)
-{
-    while (1)
-    {
-        char data[CHECKSUM_MAX_BITS + 1];
-        int data_status = safe_input_binary_string(
-            data, sizeof(data), "\n\nCRC Demo\nEnter binary data or 'X' to exit:- ");
-
-        if (data_status == INPUT_EXIT_SIGNAL)
-        {
-            printf("\nExiting CRC demo....");
-            return;
-        }
-
-        if (data_status == 0)
-        {
-            continue;
-        }
-
-        char generator[CHECKSUM_MAX_BITS + 1];
-        int generator_status = safe_input_binary_string(
-            generator, sizeof(generator), "Enter generator polynomial or 'X' to exit:- ");
-
-        if (generator_status == INPUT_EXIT_SIGNAL)
-        {
-            printf("\nExiting CRC demo....");
-            return;
-        }
-
-        if (generator_status == 0)
-        {
-            continue;
-        }
-
-        int generator_len = (int)strlen(generator);
-
-        if (generator_len < 2)
-        {
-            printf("\ngenerator polynomial must contain at least 2 bits.\n");
-            continue;
-        }
-
-        char remainder[CHECKSUM_MAX_BITS + 1];
-        char codeword[(CHECKSUM_MAX_BITS * 2) + 1];
-
-        crc_generate(data, generator, remainder, codeword);
-
-        printf("\nOriginal Data : %s", data);
-        printf("\nGenerator     : %s", generator);
-        printf("\nCRC Remainder : %s", remainder);
-        printf("\nTransmitted Codeword : %s\n", codeword);
-    }
-}
